requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
idna==3.7
msal==1.28.1
pycparser==2.22
PyJWT==2.8.0
requests==2.32.3
soupsieve==2.5
urllib3==2.2.2
flask
openai
python-dotenv
selenium
gunicorn
psycopg2
markdown
flask_wtf
